add parseType util file
